refactor: 폴더 구조 변경
- cardInfo폴더의 컴포넌트들이 많아져서 가독성을 위해 관심사별 하위폴더 생성
